requirements.txt
rich
langchain-openai
langchain
langgraph
chainlit
python-telegram-bot==13.15
